move to separate package, add details property
