--- src/main/java/com/teamchallenge/bookti/exception/ErrorResponse.java
@@ -1,23 +1,0 @@
-package com.teamchallenge.bookti.exception;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-@Builder
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ErrorResponse {
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd hh:mm:ssa")
-    private LocalDateTime timestamp;
-
-    @JsonProperty("status_code")
-    private Integer status;
-
-    private String message;
-}
